requirements.txt
python-telegram-bot
python-dotenv
pymongo
pytz
tzdata
apscheduler
python-telegram-bot[job-queue]